requirements.txt
flask==3.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
werkzeug==3.0.0